+++ app/lib/features/my_bookings/presentation/manager/my_bookings_bloc/my_bookings_event.dart
@@ -24,3 +24,7 @@ final class RefreshMyBookingsEvent extends MyBookingsEvent {
   @override
   List<Object?> get props => [clientId];
 }
+
+final class ResetMyBookingsEvent extends MyBookingsEvent {
+  const ResetMyBookingsEvent();
+}